chore: vo 패키지 변경
