+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.domain.vo.Name;
+import kitchenpos.menu.domain.vo.Name;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
