+++ src/test/java/kitchenpos/menu/domain/vo/NameTest.java
@@ -1,10 +1,10 @@
-package kitchenpos.domain.vo;
+package kitchenpos.menu.domain.vo;
 
+import kitchenpos.menu.domain.vo.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.junit.jupiter.api.Assertions.*;
 
 class NameTest {
