+++ src/test/java/kitchenpos/order/domain/OrderRepositoryTest.java
@@ -7,9 +7,9 @@ import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.menu.domain.Product;
 import kitchenpos.table.domain.TableGroup;
-import kitchenpos.domain.vo.Name;
-import kitchenpos.domain.vo.Price;
-import kitchenpos.domain.vo.Quantity;
+import kitchenpos.menu.domain.vo.Name;
+import kitchenpos.menu.domain.vo.Price;
+import kitchenpos.menu.domain.vo.Quantity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
